+++ test/integration_test.dart
@@ -21,17 +21,26 @@ main() {
         addTask(new GrinderTask('foo', taskFunction: _fooTask));
         addTask(
             new GrinderTask('bar', taskFunction: _barTask, depends: ['foo']));
+        addTask(new GrinderTask('baz', taskFunction: _bazTask));
       }
 
       _clear();
     });
 
-    grinderTest('all ran', () {
-      return runTasks(['bar']);
-    }, (ctx) {
+    grinderTest('run dependent tasks',
+        () => runTasks(['bar', '--flag', '--option=123', 'baz']), (ctx) {
       expect(ctx.isFailed, false);
+
+      // run dependent tasks
       expect(ranTasks['foo'], true);
       expect(ranTasks['bar'], true);
+
+      // pass args
+      expect(ranTasks['flag'], true);
+      expect(ranTasks['option'], '123');
+
+      // old form
+      expect(ranTasks['baz'], true);
     });
   });
 }
@@ -40,10 +49,21 @@ void _clear() => ranTasks.clear();
 
 _fooTask() {
   ranTasks['foo'] = true;
+
   log('ran _fooTask');
 }
 
-_barTask([TaskArgs args]) {
+_barTask(TaskArgs args) {
   ranTasks['bar'] = true;
+  ranTasks['flag'] = args.getFlag('flag');
+  ranTasks['option'] = args.getOption('option');
+
   log('ran _barTask\n${context}');
+}
+
+// old form
+_bazTask(GrinderContext c) {
+  ranTasks['baz'] = true;
+
+  log('baz _fooTask');
 }